requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0 